REFACTOR: models names are self explanatory. ADD: restaurants request now directly return a list of restaurants
