--- app/src/main/java/com/example/pablojeria/pruebarestapi/models/RestaurantCollections.java
@@ -1,13 +1,0 @@
-package com.example.pablojeria.pruebarestapi.models;
-
-public class RestaurantCollections {
-    private RestaurantCollectionsCollection collection;
-
-    public RestaurantCollectionsCollection getCollection() {
-        return this.collection;
-    }
-
-    public void setCollection(RestaurantCollectionsCollection collection) {
-        this.collection = collection;
-    }
-}
