+++ app/src/main/java/com/example/pablojeria/pruebarestapi/models/RestaurantContainer.java
@@ -1,0 +1,17 @@
+package com.example.pablojeria.pruebarestapi.models;
+
+public class RestaurantContainer {
+
+    private Restaurant collection;
+
+    public RestaurantContainer() {
+    }
+
+    public Restaurant getCollection() {
+        return this.collection;
+    }
+
+    public void setCollection(Restaurant collection) {
+        this.collection = collection;
+    }
+}
